Renamed IO_error to Connection_error.
