+++ include/onedatashare/credential_service.h
@@ -65,7 +65,7 @@ namespace One_data_share {
          * 
          * @return a string containing the OAuth url
          * 
-         * @exception IO_error if unable to connect to One Data Share
+         * @exception Connection_error if unable to connect to One Data Share
          * @exception Unexpected_response_error if an unexpected response
          * is received from One Data Share
          */
@@ -88,7 +88,7 @@ namespace One_data_share {
          * @param secret borrowed reference to the password needed to log in
          * to the endpoint
          * 
-         * @exception IO_error if unable to connect to One Data Share
+         * @exception Connection_error if unable to connect to One Data Share
          * @exception Unexpected_response_error if an unexpected response
          * is recieved from One Data Share
          */
