+++ src/ods_error.cpp
@@ -4,7 +4,7 @@
  * 7/16/20
  */
 
-#include <ods_error.h>
+#include <onedatashare/ods_error.h>
 
 namespace One_data_share {
     Ods_error::Ods_error(const std::string& what_arg)
@@ -15,13 +15,13 @@ namespace One_data_share {
 
     Ods_error::~Ods_error() = default;
 
-    IO_error::IO_error(const std::string& what_arg)
+    Connection_error::Connection_error(const std::string& what_arg)
         : Ods_error(what_arg) {}
 
-    IO_error::IO_error(const char* what_arg)
+    Connection_error::Connection_error(const char* what_arg)
         : Ods_error(what_arg) {}
 
-    IO_error::~IO_error() = default;
+    Connection_error::~Connection_error() = default;
 
     Unexpected_response_error::Unexpected_response_error(const std::string& what_arg, int status)
         : Ods_error(what_arg), status(status) {}
