+++ src/rest.h
@@ -87,7 +87,7 @@ namespace One_data_share {
 		 * @return the Response object containing the response headers,
 		 * body, and http status
 		 * 
-		 * @exception IO_error if unable to connect to the sepcified
+		 * @exception Connection_error if unable to connect to the sepcified
 		 * url
 		 */
 		virtual Response get(const std::string& url, const std::unordered_multimap<std::string, std::string>& headers) const = 0;
@@ -107,7 +107,7 @@ namespace One_data_share {
 		 * @return the Response object containing the response headers,
 		 * body, and http status
 		 * 
-		 * @exception IO_error if unable to connect to the sepcified
+		 * @exception Connection_error if unable to connect to the sepcified
 		 * url
 		 */
 		virtual Response post(const std::string& url, const std::unordered_multimap<std::string, std::string>& headers, const std::string& data) const = 0;
